backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
python-jose>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9
razorpay==2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
